add basic ffi

+++ runtime/files/allocation.c
@@ -5,7 +5,7 @@
 
 #include "gc.h"
 
-int * allocateArray(int size) {
-    int * x = GC_MALLOC(size * 8);
+uint64_t * allocateArray(int size) {
+    uint64_t * x = GC_MALLOC(size * 8);
     return x;
 }
